DEV-3 regex Compile to MustCompile, error lowrcase

+++ src/DataTransfer/podDataCreator/domain/measurement/value/number/Number.go
@@ -3,7 +3,7 @@ package number
 import "fmt"
 
 type Number struct {
-	value        float32
+	value        float64
 	podUnits     Unit
 	displayUnits Unit
 }
@@ -15,7 +15,7 @@ func NewNumber(podUnitString string, displayUnitString string) *Number {
 }
 
 func (i *Number) ToDisplayString() string {
-	number := float32(i.value)
+	number := float64(i.value)
 	internationalSystemNumber := undoUnits(number, i.podUnits.operations)
 	result := convertToUnits(internationalSystemNumber, i.displayUnits.operations)
 	return fmt.Sprintf("%v", result)
